Split grpc and grpcio examples
